sigref/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-chartjs==2.3.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.15.2
drf-nested-routers==0.94.1
Faker==37.0.0
gunicorn==23.0.0
jsonfield==3.1.0
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
psycopg2-binary==2.9.10
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyserial==3.5
pytz==2025.2
pywin32-ctypes==0.2.3
setuptools==76.0.0
sqlparse==0.5.3
swapper==1.4.0
typing_extensions==4.13.0
tzdata==2025.1
whitenoise==6.9.0
